+++ NEWS.md
@@ -1,3 +1,13 @@
+# [urbanAnnualRunoff 0.1.0](https://github.com/KWB-R/urbanAnnualRunoff/tag/v0.1.0) <small>2022-05-18</small>
+
+The workflow for modelling urban annual runoff using the water balance model [ABIMO](https://github.com/umweltatlas/abimo) is described for the 
+two case-study sites in China here: 
+
+- [Jinxi](https://kwb-r.github.io/urbanAnnualRunoff/articles/workflow_jinxi.html) and 
+
+- [Tongzhou (Beijing)](https://kwb-r.github.io/urbanAnnualRunoff/articles/workflow_beijing.html)
+
+
 # urbanAnnualRunoff 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.